Don't initialise particle indices to -1
There doesn't seem to be any point to initialising these to a sentinel
value anyway. If a particle is on the very leftmost or bottommost edge
of the domain, the logic in search_indices() dictates that the index
shouldn't be incremented. In this case, the index remains at -1,
and xsi accidentally ends up being somewhat reasonable.

If the indices are instead just initialised to 0, the search starts by
assuming the particle is in the first cell, and immediately terminates
in the above condition, which is what we'd expect.

Fixes https://github.com/angus-g/lagrangian-filtering/issues/10