Fix Kernel cleanup
The Kernel class tried to clean up the source, log and library files
that were created during JIT compilation. However, on interpreter
exit, Python would clean up the modules required in the destructor
before the destructor was actually called -- meaning there'd be an
exception when the interpreter quits. We move instead to a
tempfile.TemporaryDirector, and check that we can unload the shared
library when the Kernel is destroyed.